Phase 4: GPU batch parallel + framebuffer optimization
- NESBatchGpu class: run N NES instances in parallel (1 thread/block)
- 4 batch CUDA kernels: batch_run_frame, batch_reset, batch_step_frames, batch_get_framebuffers
- 8/8 GPU batch tests passing
- Framebuffer optimization: uint32_t[61440] -> uint8_t[61440] (palette index storage)
  - NESState: 244KB -> 64KB/instance (3.8x reduction)
  - Palette-to-RGBA conversion via NES_PALETTE_CONST on output only
  - All 18/18 GPU tests passing after optimization
- Benchmark results:
  - 1000 instances: 25-27k SPS (99-107x vs nes_py)
  - 2000 instances: 41-43k SPS (164-171x)  <- 120x target exceeded
  - 5000 instances: 48-49k SPS (191-196x)
  - Peak ~50k SPS (198x) -- 65% above 120x target
- Updated CURRENT_STATUS.md: Phase 4 complete, 83% overall progress

+++ nes_emulator_gpu/src/cuda/device/nes_state.h
@@ -98,8 +98,10 @@ struct NESPPUState {
     uint8_t  palette[NES_PALETTE_SIZE];     // 32-byte palette RAM
     uint8_t  oam[NES_OAM_SIZE];             // 256-byte Object Attribute Memory
 
-    // ---- Output framebuffer (256x240 RGBA8888) ----
-    uint32_t framebuffer[NES_FRAMEBUFFER_SIZE];
+    // ---- Output framebuffer (256×240, palette index 0-63 per pixel) ----
+    // Stores NES palette index (uint8_t) to minimize state size (60KB vs 240KB).
+    // Use nes_get_framebuffer / nes_batch_get_framebuffers to get RGBA32 output.
+    uint8_t  framebuffer[NES_FRAMEBUFFER_SIZE];
 
     // ---- Sprite rendering state (rebuilt each scanline) ----
     ActiveSpriteGPU active_sprites[NES_MAX_SPRITES];
